popitka 3/Koshelok3/buckend/requirements.txt
requests==2.31.0
fastapi==0.96.0
pyjwt==2.7.0
uvicorn==0.22.0
peewee==3.17.1
psycopg2==2.9.9
sqlalchemy==2.0.30
tonsdk==1.0.13
pytonconnect==0.3.1
cryptography==42.0.7